refactor: Move `getBundleDTO` from `BundleController` to `BundleService`
- Extracted `getBundleDTO` from `BundleController` to `BundleService`.
